perftests: split out files into utils and tests, rename files
